+++ docs/demo/panel.md
@@ -1,0 +1,8 @@
+---
+title: Panel
+nav:
+  title: Demo
+  path: /demo
+---
+
+<code src="../../examples/panel.tsx"></code>
